refactor: move `server_definitions` code to its own files
